// sanity/schemaTypes/objects/technologies-used/types.ts
export enum TechnologiesFields {
  TechnologyChoice = 'technologyChoice',
  FreeText = 'freeText'
}


export type TechnologyChoice = {
  technologyChoice?: string;
  freeText?: string;
}


export enum Technology {
  JavaScript = 'javascript',
  TypeScript = 'typescript',
  Python = 'python',
  Java = 'java',
  CSharp = 'csharp',
  CPlusPlus = 'cplusplus',
  Go = 'go',
  Ruby = 'ruby',
  PHP = 'php',
  Swift = 'swift',
  Kotlin = 'kotlin',
  Rust = 'rust',
  Dart = 'dart',
  Scala = 'scala',
  ObjectiveC = 'objective-c',
  Shell = 'shell',
  Bash = 'bash',
  Perl = 'perl',
  R = 'r',
  MATLAB = 'matlab',
  SQL = 'sql',
  GraphQL = 'graphql',
  HTML = 'html',
  CSS = 'css',
  Sass = 'sass',
  Less = 'less',
  React = 'react',
  NextJS = 'nextjs',
  Vue = 'vue',
  Nuxt = 'nuxt',
  Angular = 'angular',
  Svelte = 'svelte',
  NodeJS = 'nodejs',
  Express = 'express',
  NestJS = 'nestjs',
  Django = 'django',
  Flask = 'flask',
  FastAPI = 'fastapi',
  Spring = 'spring',
  Laravel = 'laravel',
  Rails = 'rails',
  '.NET' = 'dotnet',
  MongoDB = 'mongodb',
  PostgreSQL = 'postgresql',
  MySQL = 'mysql',
  Redis = 'redis',
  Docker = 'docker',
  Kubernetes = 'kubernetes',
  AWS = 'aws',
  Azure = 'azure',
  GCP = 'gcp',
  Firebase = 'firebase',
  Jenkins = 'jenkins',
  Git = 'git',
  Github = 'github',
  Gitlab = 'gitlab',
  Bitbucket = 'bitbucket',
  Jira = 'jira',
  Figma = 'figma',
  Storybook = 'storybook',
  Cypress = 'cypress',
  Jest = 'jest',
  Mocha = 'mocha',
  Chai = 'chai',
  Puppeteer = 'puppeteer',
  Playwright = 'playwright',
  Webpack = 'webpack',
  Babel = 'babel',
  ESLint = 'eslint',
  Prettier = 'prettier',
  Tailwind = 'tailwind',
  Bootstrap = 'bootstrap',
  MaterialUI = 'materialui',
  AntDesign = 'antdesign',
  ChakraUI = 'chakraui',
  ThreeJS = 'threejs',
  D3 = 'd3',
  ChartJS = 'chartjs',
  Redux = 'redux',
  MobX = 'mobx',
  Zustand = 'zustand',
  Recoil = 'recoil',
  RxJS = 'rxjs',
  Apollo = 'apollo',
  Prisma = 'prisma',
  Sequelize = 'sequelize',
  Mongoose = 'mongoose',
  Electron = 'electron',
  ReactNative = 'reactnative',
  Expo = 'expo',
  Cordova = 'cordova',
  Capacitor = 'capacitor',
  Unity = 'unity',
  Unreal = 'unreal',
  Blender = 'blender',
  WebAssembly = 'webassembly',
  Solidity = 'solidity',
  Hardhat = 'hardhat',
  Truffle = 'truffle',
  Vite = 'vite',
  Gatsy = 'gatsby',
  Sanity = 'sanity',
  Contentful = 'contentful',
  Strapi = 'strapi',
  WordPress = 'wordpress',
  Shopify = 'shopify',
  Magento = 'magento',
  WooCommerce = 'woocommerce',
  Squarespace = 'squarespace',
  Wix = 'wix',
  Netlify = 'netlify',
  Vercel = 'vercel',
  Heroku = 'heroku',
  DigitalOcean = 'digitalocean',
  Linode = 'linode',
  Cloudflare = 'cloudflare',
  Sentry = 'sentry',
  Datadog = 'datadog',
  NewRelic = 'newrelic',
  Segment = 'segment',
  Mixpanel = 'mixpanel',
  Amplitude = 'amplitude',
  Hotjar = 'hotjar',
  GoogleAnalytics = 'googleanalytics',
  Mailchimp = 'mailchimp',
  Sendgrid = 'sendgrid',
  Twilio = 'twilio',
  Stripe = 'stripe',
  Paypal = 'paypal',
  Plaid = 'plaid',
  Algolia = 'algolia',
  ElasticSearch = 'elasticsearch',
  RabbitMQ = 'rabbitmq',
  Kafka = 'kafka',
  Nginx = 'nginx',
  Apache = 'apache',
  Terraform = 'terraform',
  Ansible = 'ansible',
  Chef = 'chef',
  Puppet = 'puppet',
  Salt = 'salt',
  Prometheus = 'prometheus',
  Grafana = 'grafana'
}